backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pgvector
pydantic
python-multipart
python-docx
reportlab
google-generativeai